+++ app/src/main/java/top/liplus/v4over6/vpn/Statistics.java
@@ -1,8 +1,8 @@
 package top.liplus.v4over6.vpn;
 
 public class Statistics {
-    public int uploadPackets = 0;
-    public int uploadBytes = 0;
-    public int downloadPackets = 0;
-    public int downloadBytes = 0;
+    public long uploadPackets = 0;
+    public long uploadBytes = 0;
+    public long downloadPackets = 0;
+    public long downloadBytes = 0;
 }
